+++ src/components/AlteracaoDeCardapio/Terceirizada/PainelPedidos/Container.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { visaoPorComboSomenteDatas } from "../../../../constants/painelPedidos.constants";
 import {
-  getTerceirizadaPedidosAprovados as pedidosAprovadosNormais,
+  getTerceirizadaPedidosAutorizados as pedidosAutorizadosNormais,
   getTerceirizadaPedidosReprovados as pedidosReprovadosNormais
 } from "../../../../services/alteracaoDecardapio.service";
 import PainelPedidos from ".";
@@ -11,18 +11,18 @@ class Container extends Component {
     super(props);
     this.state = {
       visaoPorCombo: visaoPorComboSomenteDatas,
-      pedidosAprovados: [],
+      pedidosAutorizados: [],
       pedidosReprovados: []
     };
   }
 
   componentDidMount() {
-    let pedidosAprovados = [];
+    let pedidosAutorizados = [];
     let pedidosReprovados = [];
 
-    pedidosAprovadosNormais().then(response => {
-      pedidosAprovados = response.results;
-      this.setState({ pedidosAprovados });
+    pedidosAutorizadosNormais().then(response => {
+      pedidosAutorizados = response.results;
+      this.setState({ pedidosAutorizados });
     });
 
     pedidosReprovadosNormais().then(response => {
