+++ src/components/AlteracaoDeCardapio/components/CardPendenteAcao/style.scss
