+++ src/components/InclusaoDeAlimentacao/components/CardPendenteAcao/style.scss
