+++ src/components/InversaoDeDiaDeCardapio/CODAE/PainelPedidos/index.jsx
@@ -3,7 +3,7 @@ import { connect } from "react-redux";
 import { Field, formValueSelector, reduxForm } from "redux-form";
 import { FiltroEnum } from "../../../../constants/filtroEnum";
 import { getCODAEPedidosDeInversoes } from "../../../../services/inversaoDeDiaDeCardapio.service";
-import { CardInversaoPendenciaAprovacao } from "../../components/CardPendenciaAprovacao";
+import { CardInversaoPendenciaAprovacao } from "../../components/CardPendenteAcao";
 import { TIPODECARD } from "../../../../constants/cardsPrazo.constants";
 import CardHistorico from "../../components/CardHistorico";
 import { filtraPrioritarios } from "../../../../helpers/painelPedidos";
