+++ src/components/InversaoDeDiaDeCardapio/components/CardPendenteAcao/style.scss
