+++ src/components/SolicitacaoUnificada/components/CardPendenteAcao/index.jsx
@@ -1,25 +1,34 @@
 import React, { Component } from "react";
 import { Collapse } from "react-collapse";
 import { Link } from "react-router-dom";
-import { calcularNumeroDeEscolasUnicas } from "./helper";
 import { talvezPluralizar } from "../../../../helpers/utilities";
+import { calcularNumeroDeDREsUnicas } from "./helper";
 import "./style.scss";
-import { SOLICITACAO_KIT_LANCHE } from "../../../../configs/constants";
+import { SOLICITACAO_KIT_LANCHE_UNIFICADA } from "../../../../configs/constants";
 import { ToggleExpandir } from "../../../Shareable/ToggleExpandir";
 
-export class CardPendenciaAprovacao extends Component {
+export const TIPO_CARD_ENUM = {
+  LIMITE: "on-limit",
+  REGULAR: "regular",
+  PRIORIDADE: "priority"
+};
+
+export class CardPendenteAcao extends Component {
   constructor(props) {
     super(props);
     this.state = {
       collapsed: true,
-      pedidosFiltrados: this.props.pedidos
+      pedidosFiltrados: [],
+      filtrado: false
     };
     this.filtrarPedidos = this.filtrarPedidos.bind(this);
   }
 
-  componentDidUpdate(prevProps) {
-    if (this.props.pedidos.length !== prevProps.pedidos.length) {
-      this.setState({ pedidosFiltrados: this.props.pedidos });
+  componentDidUpdate(prevProps, nextProps) {
+    if (this.props.pedidos.length !== nextProps.pedidosFiltrados.length) {
+      if (this.state.filtrado === false) {
+        this.setState({ pedidosFiltrados: this.props.pedidos });
+      }
     }
   }
 
@@ -29,12 +38,13 @@ export class CardPendenciaAprovacao extends Component {
     pedidosFiltrados = pedidosFiltrados.filter(function(item) {
       const palavraAFiltrar = event.target.value.toLowerCase();
       return (
-        item.id_externo.toLowerCase().includes(palavraAFiltrar) ||
-        item.escola.nome.toLowerCase().search(palavraAFiltrar) !== -1 ||
-        item.escola.codigo_eol.includes(palavraAFiltrar)
+        item.id_externo.toLowerCase().search(palavraAFiltrar) !== -1 ||
+        item.diretoria_regional.nome.toLowerCase().search(palavraAFiltrar) !==
+          -1
+        //item.lote.nome.includes(palavraAFiltrar)
       );
     });
-    this.setState({ pedidosFiltrados });
+    this.setState({ pedidosFiltrados, filtrado: true });
   }
 
   render() {
@@ -47,7 +57,7 @@ export class CardPendenciaAprovacao extends Component {
     } = this.props;
     const { collapsed, pedidosFiltrados } = this.state;
     return (
-      <div className="card card-pendency-approval meal-kit-solicitation">
+      <div className="card card-pendency-approval unified-solicitation">
         <div className={"card-title " + tipoDeCard}>{titulo}</div>
         <div className="row">
           <div className="col-2">
@@ -64,14 +74,14 @@ export class CardPendenciaAprovacao extends Component {
                 <div className="label" />
                 <span className="text">
                   <span className="value">
-                    {calcularNumeroDeEscolasUnicas(pedidos)}{" "}
+                    {calcularNumeroDeDREsUnicas(pedidos)}{" "}
                   </span>
                   {`
                   ${talvezPluralizar(
-                    calcularNumeroDeEscolasUnicas(pedidos),
+                    calcularNumeroDeDREsUnicas(pedidos),
                     "escola"
                   )} ${talvezPluralizar(
-                    calcularNumeroDeEscolasUnicas(pedidos),
+                    calcularNumeroDeDREsUnicas(pedidos),
                     "solicitante"
                   )}
                   `}
@@ -99,13 +109,13 @@ export class CardPendenciaAprovacao extends Component {
               />
               <i className="fas fa-search inside-input" />
             </div>
-            <table className="orders-table mt-4 ml-3 mr-3">
+            <table className="orders-table-unified-solicitation mt-4 mr-3">
               <thead>
-                <tr>
-                  <th>Código do Pedido</th>
-                  <th>Código EOL</th>
-                  <th>Nome da Escola</th>
-                  <th>{ultimaColunaLabel || "Data"}</th>
+                <tr className="row">
+                  <th className="col-3">Código do Pedido</th>
+                  <th className="col-3">Lote</th>
+                  <th className="col-3">DRE</th>
+                  <th className="col-3">{ultimaColunaLabel || "Data"}</th>
                 </tr>
               </thead>
               <tbody>
@@ -114,15 +124,22 @@ export class CardPendenciaAprovacao extends Component {
                     return (
                       <Link
                         key={key}
-                        to={`/${parametroURL}/${SOLICITACAO_KIT_LANCHE}/relatorio?uuid=${
+                        to={`/${parametroURL}/${SOLICITACAO_KIT_LANCHE_UNIFICADA}/relatorio?uuid=${
                           pedido.uuid
                         }`}
                       >
-                        <tr>
-                          <td>{pedido.id_externo}</td>
-                          <td>{pedido.escola.codigo_eol}</td>
-                          <td>{pedido.escola.nome}</td>
-                          <td>{pedido.solicitacao_kit_lanche.data}</td>
+                        <tr className="row">
+                          <td className="col-3">{pedido.id_externo}</td>
+                          <td className="col-3">
+                            {pedido.escolas_quantidades[0] &&
+                              pedido.escolas_quantidades[0].escola.lote.nome}
+                          </td>
+                          <td className="col-3">
+                            {pedido.diretoria_regional.nome}
+                          </td>
+                          <td className="col-3">
+                            {pedido.solicitacao_kit_lanche.data}
+                          </td>
                         </tr>
                       </Link>
                     );
