+++ src/components/SuspensaoDeAlimentacao/components/CardPendenteAcao/style.scss
